+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.2.4 - 2015-11-28
 * Get projects.json from firebase instead of divshot (issue #7)
 * Use the new logo in the splash screen
+* Add data (last commit, daily star variation) in the `<ProjectPage>`
 
 ## 0.2.3 - 2015-11-08
 * New sidebar using [slideout](https://github.com/Mango/slideout) JavaScript component instead of pure CSS.